requirements.txt
click==8.1.3
Flask==2.2.2
Flask-Login==0.6.2
Flask-MySQLdb==1.0.1
Flask-WTF==1.0.1
gunicorn==20.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
mysqlclient==2.1.1
psycopg2==2.9.5
python-dotenv==0.21.1
Werkzeug==2.2.2
WTForms==3.0.1
